config/main.py
import logging, os
from dotenv import load_dotenv

load_dotenv()

logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

logger = logging.getLogger(__name__)

# logger.info("+")

db_url = os.getenv('DATABASE_URL')
