requirements.txt
sentence-transformers==3.4.1
pymongo==4.11.2
python-dotenv==1.0.1